burrego: move cli program to an examples
Burrego is just a library, this commit moves the main program under a
Rust example. This makes possible to remove the `clap` dependency from
the Burrego library.

In addition to that, the CLI program is defining its cli interface by
using the new derive-based approach of `clap`.

This fixes all the clippy warnings raised by clap.
