+++ crates/burrego/Cargo.toml
@@ -9,7 +9,6 @@ edition = "2018"
 [dependencies]
 anyhow = "1.0.57"
 base64 = "0.13.0"
-clap = { version = "3.0.14", features = [ "cargo", "env" ] }
 chrono = "0.4.19"
 chrono-tz = "0.6.1"
 gtmpl = "0.7.1"
@@ -29,3 +28,4 @@ wasmtime = "0.34.0"
 
 [dev-dependencies]
 assert-json-diff = "2.0.1"
+clap = { version = "3.1", features = [ "derive" ] }
